Remove MarketDiscovery module
