--- UnstoppableWallet/UnstoppableWallet/Modules/Market/MarketDiscovery/MarketDiscoveryFilterService.swift
@@ -1,94 +1,0 @@
-import Foundation
-import MarketKit
-import RxRelay
-import RxSwift
-
-class MarketDiscoveryFilterService {
-    private let marketKit: MarketKit.Kit
-    private let favoritesManager: FavoritesManager
-
-    private let stateRelay = PublishRelay<State>()
-    private(set) var state: State = .idle {
-        didSet {
-            stateRelay.accept(state)
-        }
-    }
-
-    private let resultRelay = PublishRelay<Result>()
-
-    init(marketKit: MarketKit.Kit, favoritesManager: FavoritesManager) {
-        self.marketKit = marketKit
-        self.favoritesManager = favoritesManager
-    }
-
-    private func coinUid(index: Int) -> String? {
-        guard case let .searchResults(fullCoins) = state, index < fullCoins.count else {
-            return nil
-        }
-
-        return fullCoins[index].coin.uid
-    }
-}
-
-extension MarketDiscoveryFilterService {
-    var stateObservable: Observable<State> {
-        stateRelay.asObservable()
-    }
-
-    var resultObservable: Observable<Result> {
-        resultRelay.asObservable()
-    }
-
-    func set(filter: String) {
-        if filter.isEmpty {
-            state = .idle
-        } else {
-            do {
-                state = try .searchResults(fullCoins: marketKit.fullCoins(filter: filter))
-            } catch {
-                state = .searchResults(fullCoins: [])
-            }
-        }
-    }
-
-    func isFavorite(index: Int) -> Bool {
-        guard let coinUid = coinUid(index: index) else {
-            return false
-        }
-
-        return favoritesManager.isFavorite(coinUid: coinUid)
-    }
-
-    func favorite(index: Int) {
-        guard let coinUid = coinUid(index: index) else {
-            resultRelay.accept(.fail)
-            return
-        }
-
-        favoritesManager.add(coinUid: coinUid)
-        resultRelay.accept(.favorited)
-    }
-
-    func unfavorite(index: Int) {
-        guard let coinUid = coinUid(index: index) else {
-            resultRelay.accept(.fail)
-            return
-        }
-
-        favoritesManager.remove(coinUid: coinUid)
-        resultRelay.accept(.unfavorited)
-    }
-}
-
-extension MarketDiscoveryFilterService {
-    enum State {
-        case idle
-        case searchResults(fullCoins: [FullCoin])
-    }
-
-    enum Result {
-        case favorited
-        case unfavorited
-        case fail
-    }
-}
