--- UnstoppableWallet/UnstoppableWallet/Modules/Market/MarketDiscovery/MarketDiscoveryTitleCell.swift
@@ -1,33 +1,0 @@
-import UIKit
-
-class MarketDiscoveryTitleCell: UICollectionViewCell {
-    private let label = UILabel()
-    private let topSeparatorView = UIView()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        contentView.addSubview(label)
-        label.snp.makeConstraints { maker in
-            maker.leading.trailing.equalToSuperview().inset(CGFloat.margin16)
-            maker.centerY.equalToSuperview()
-        }
-
-        label.font = .body
-        label.textColor = .themeLeah
-        label.text = "market_discovery.browse_categories".localized
-
-        contentView.addSubview(topSeparatorView)
-        topSeparatorView.snp.makeConstraints { maker in
-            maker.leading.top.trailing.equalToSuperview()
-            maker.height.equalTo(CGFloat.heightOnePixel)
-        }
-
-        topSeparatorView.backgroundColor = .themeSteel20
-    }
-
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
